model card utils
