+++ bpy_speckle/connector/utils/model_card_utils.py
@@ -1,0 +1,70 @@
+import bpy
+from bpy.types import Context
+from ..ui.model_card import speckle_model_card
+
+
+def update_model_card_objects(model_card, converted_objects):
+    for obj in converted_objects.values():
+        # if its a collection, add it to collections field of model card
+        if isinstance(obj, bpy.types.Collection):
+            if obj.name in (o.name for o in model_card.collections):
+                continue
+            s_col = model_card.collections.add()
+            s_col.name = obj.name
+        # if its an object, add it to the objects field of model card
+        if isinstance(obj, bpy.types.Object):
+            if obj.name in (o.name for o in model_card.objects):
+                continue
+            s_obj = model_card.objects.add()
+            s_obj.name = obj.name
+
+
+def delete_model_card_objects(model_card: speckle_model_card, context: Context) -> None:
+    """
+    deletes the model card objects
+    """
+    select_model_card_objects(model_card, context)
+    bpy.ops.object.delete()
+    # delete model card/currently loaded collections
+    for col in model_card.collections:
+        coll = bpy.data.collections.get(col.name)
+        if not coll:
+            continue
+        # unlink from scenes
+        for scene in bpy.data.scenes:
+            if coll.name in scene.collection.children:
+                scene.collection.children.unlink(coll)
+        bpy.data.collections.remove(coll)
+
+
+def select_model_card_objects(model_card, context: Context):
+    # deselect all objects first
+    bpy.ops.object.select_all(action="DESELECT")
+    # select objects in model card
+    for obj in model_card.objects:
+        blender_obj = bpy.data.objects.get(obj.name)
+        if not blender_obj:
+            continue
+        if blender_obj.name in context.view_layer.objects:
+            blender_obj.select_set(True)
+
+    selected = context.selected_objects
+    if selected:
+        context.view_layer.objects.active = selected[0]
+
+
+def zoom_to_selected_objects(context: Context):
+    """
+    zooms to the selected objects
+    """
+    bpy.ops.view3d.view_selected()
+
+
+def model_card_exists(project_id: str, model_id: str, context: Context) -> bool:
+    """
+    checks if a model card exists
+    """
+    for model_card in context.scene.speckle_state.model_cards:
+        if model_card.project_id == project_id and model_card.model_id == model_id:
+            return True
+    return False
